Add support for files larger than 2GB (See SR #1583). - Replace PHP functions that don't work on files larger than 2GB with workarounds (mainly system calls). Some workarounds could be suppressed in a future version when PHP is able to handle files larger that 2 GB. - DB structure is also modified (the number for a file size of 2GB  cannot be stored in normal integer, so use bigint). - Updated SELinux rules to allow corresponding system calls. Tested on 3 GB file.
Some limitations:
- SOAP calls to FRS service don't handle 2GB+ files
- only FRS is fixed: documents, tracker attachments and wiki attachments cannot be larger than 2GB
- workarounds are located in file_utils.php (and FRSFile.class.php): this could be improved.

+++ src/db/mysql/database_structure.sql
@@ -782,7 +782,7 @@ CREATE TABLE frs_file (
   type_id int(11) NOT NULL default '0',
   processor_id int(11) NOT NULL default '0',
   release_time int(11) NOT NULL default '0',
-  file_size int(11) NOT NULL default '0',
+  file_size bigint NOT NULL default '0',
   post_date int(11) NOT NULL default '0',
   status char(1) NOT NULL default 'A',
   PRIMARY KEY  (file_id),
@@ -2949,7 +2949,7 @@ CREATE TABLE wiki_attachment_revision (
   date INT( 11 ) NOT NULL ,
   revision INT( 11 ) NOT NULL ,
   mimetype VARCHAR( 255 ) NOT NULL ,
-  size INT( 11 ) NOT NULL ,
+  size bigint NOT NULL ,
   PRIMARY KEY (id)
 );
